Fix bug in version computation
As pointed out in #2, version computation was using grid size as if
it were the version itself. this meant that grids larger than 40 modules
were rejected.

This was fixed by first computing the version used on the module
heuristic and only then checking if the value is in the expected range.