fix undetected capstones at image edges
The line scanner triggered a search for capstones after seeing
5 black/white transitions. If a capstone lies right on the edge of an
image, it would never trigger the a full check. To remedy this, we insert
a virtual white pixel at the end of a line. This should trigger the line
scanner to find the capstone.

Adds a basic unit test that triggers the behaviour.