deal with out-of-bounds access in slanted grids
Our grid transformation allows for buffer coordinates that are out of
bounds. This would lead to panics when checking modules in these
locations. The fix is to clamp the possible coordinate values so they
always point to in-bound coordinates.

Includes a slated image that triggered the panic in the old code.