Drop unused dependencies and most test data from published crates
This reduces the download size for the rqrr crate from crates.io
from almost 3 MB to about 40 kB (1-2% of the current size).

The "criterion" dependency is currently unused since all code in
benches/* is commented out, so this comments out the criterion
dependency too since it just adds compilation time for dozens of
crates that are unused.