support mirrored images
Seems like more and more QR Codes in the wild are not standard compliant and
are "mirrored". This would cause rqrr to read the wrong data.

Since the QR code standard has no way to mark the correct orientation for
mirrored codes, we just have to guess the right orientation when reading, i.e.
trying both and seeing where we get a result.